xhs-tryon/pyproject.toml
[project]
name = "jobson-xhs-mcp"
version = "0.1.1"
description = "小红书MCP服务器 - 提供小红书API功能的MCP工具"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "curl-cffi>=0.10.0,<0.14.0",
    "mcp[cli]>=1.6.0",
    "Pillow>=11.0.0",
    "playwright>=1.58.0",
    "pyexecjs>=1.5.1",
    "requests>=2.32.3",
]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
keywords = ["mcp", "xiaohongshu", "api", "server"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
fashn-tryon = "fashn_tryon.cli:main"
xhs-mcp = "xhs_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/jobsonlook/xhs-mcp"
Repository = "https://github.com/jobsonlook/xhs-mcp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fashn_tryon", "xhs_mcp"]
